IntellIJ, File Movement and new abstraction
Now, not dealing with gpus

+++ Python/MFCC/MFCCAlt.py
@@ -1,4 +1,4 @@
-#MFCCAlt.py
+# MFCCAlt.py
 
 fileLoc = '/Users/varunnayyar/Documents/Speech Stuff/LDC Cleaned/Bored/BoredCC1_1.wav'
 
@@ -7,12 +7,10 @@ from features.base import mfcc
 import scipy.io.wavfile as wav
 import numpy as np
 
-(rate,sig) = wav.read(fileLoc)
+(rate, sig) = wav.read(fileLoc)
 
-
-
-sig=sig.astype(np.float32)
-sig= 0.5*(sig.T[0]+sig.T[1])
+sig = sig.astype(np.float32)
+sig = 0.5 * (sig.T[0] + sig.T[1])
 
 # sig*=5
 # print "Sig.dtype: ", sig.dtype
@@ -20,7 +18,7 @@ sig= 0.5*(sig.T[0]+sig.T[1])
 # print "SigLen: ", len(sig)
 
 
-mfcc_feat = mfcc(sig,rate, winlen = 0.02, winstep = 0.01, ceplifter = 22, appendEnergy = True)
+mfcc_feat = mfcc(sig, rate, winlen=0.02, winstep=0.01, ceplifter=22, appendEnergy=True)
 # fbank_feat = logfbank(sig,rate)
 
 print "Mfcc.shape: ", mfcc_feat.shape
